requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-milvus
langchain-mongodb
langchain-openai
pymongo
pymilvus
langchain-huggingface
docx2txt
openpyxl
pypdf
pandas
streamlit
requests
python-dotenv
tiktoken
sentence-transformers
dashscope
